install quick start: Rename to "Setup Guide"
We previously renamed the ID/file name of this guide to
article[-_]installation.* (2e7443c4). That is consistent with SLES,
SLED, etc., but it does not reflect the content as far as I can tell.
Which is:
* an overview of the product
* instructions for important initial settings

Now, I am not sure whether "Setup Guide" is the best description of
that but it seems better than "(Installation) Quick Start" when the
installation is not mentioned.